tests/submission/PhilipTaylor/canvas/2d.clearRect.negative.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.clearRect.negative</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.clearRect.html">clearRect</a>.negative</h1>
<p class="desc">clearRect of negative sizes works</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.2d.clearRect">2d.clearRect</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="clear-100x50.png">
</p><ul id="d"></ul>
<script>
var t = async_test("clearRect of negative sizes works");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 100, 50);
ctx.clearRect(0, 0, 50, 25);
ctx.clearRect(100, 0, -50, 25);
ctx.clearRect(0, 50, 50, -25);
ctx.clearRect(100, 50, -50, -25);
_assertPixel(canvas, 25,12, 0,0,0,0, "25,12", "0,0,0,0");
_assertPixel(canvas, 75,12, 0,0,0,0, "75,12", "0,0,0,0");
_assertPixel(canvas, 25,37, 0,0,0,0, "25,37", "0,0,0,0");
_assertPixel(canvas, 75,37, 0,0,0,0, "75,37", "0,0,0,0");


});
</script>

<div id="log"></div>
</body></html>
